Separados los archivos de forma definitiva. Finalizado el makefile. Ahora el programa se ejecuta correctamente.

+++ Tarea1/producto.h
@@ -3,5 +3,5 @@
 typedef struct Producto{
     char Nombre[100];
     int Peso;
-    int Complejidad;
+    double Complejidad;
 }Producto;
